+++ jobportal/jobportal/src/main/java/com/jobportal/jobportal/configs/SecurityConfig.java
@@ -1,7 +1,6 @@
-package com.jobportal.jobportal.config;
+package com.jobportal.jobportal.configs;
 
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.web.SecurityFilterChain;
 
 @Configuration
 public class SecurityConfig {
