--- jobportal/jobportal/src/main/java/com/jobportal/jobportal/controller/OfferController.java
@@ -1,10 +1,0 @@
-package com.jobportal.jobportal.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class OfferController {
-
-
-
-}
