+++ jobportal/jobportal/src/main/java/com/jobportal/jobportal/controllers/OfferController.java
@@ -1,0 +1,23 @@
+package com.jobportal.jobportal.controllers;
+
+import com.jobportal.jobportal.repositories.OfferRepository;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/offer/")
+public class OfferController {
+
+    private final OfferRepository offerRepository;
+
+    public OfferController(OfferRepository offerRepository) {
+        this.offerRepository = offerRepository;
+    }
+
+    @GetMapping
+    public ResponseEntity<?> getOffers(){
+        return null;
+    }
+}
