+++ jobportal/jobportal/src/main/java/com/jobportal/jobportal/entities/Application.java
@@ -1,7 +1,6 @@
-package com.jobportal.jobportal.entity;
+package com.jobportal.jobportal.entities;
 
-import com.jobportal.jobportal.entity.user.Candidate;
-import com.jobportal.jobportal.entity.user.Company;
+import com.jobportal.jobportal.entities.user.Candidate;
 import jakarta.persistence.*;
 import lombok.Data;
