+++ jobportal/jobportal/src/main/java/com/jobportal/jobportal/entities/Payment.java
@@ -1,7 +1,7 @@
-package com.jobportal.jobportal.entity;
+package com.jobportal.jobportal.entities;
 
-import com.jobportal.jobportal.entity.offer.Offer;
-import com.jobportal.jobportal.entity.user.Company;
+import com.jobportal.jobportal.entities.offer.Offer;
+import com.jobportal.jobportal.entities.user.Company;
 import jakarta.persistence.*;
 import lombok.Data;
