+++ jobportal/jobportal/src/main/java/com/jobportal/jobportal/entities/offer/OfferEmploymentType.java
@@ -1,4 +1,4 @@
-package com.jobportal.jobportal.entity.offer;
+package com.jobportal.jobportal.entities.offer;
 
 import jakarta.persistence.*;
 import lombok.Data;
