+++ jobportal/jobportal/src/main/java/com/jobportal/jobportal/entities/offer/Technology.java
@@ -1,4 +1,4 @@
-package com.jobportal.jobportal.entity.offer;
+package com.jobportal.jobportal.entities.offer;
 
 import jakarta.persistence.*;
 import lombok.Data;
@@ -8,9 +8,9 @@ import java.util.HashSet;
 import java.util.Set;
 
 @Entity
-@Table(name = "employment_type")
+@Table(name = "technology")
 @Data
-public class EmploymentType {
+public class Technology {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -20,6 +20,7 @@ public class EmploymentType {
     @Column(name = "name", nullable = false)
     private String name;
 
-    @OneToMany(mappedBy = "employmentType", cascade = CascadeType.ALL, orphanRemoval = true)
-    private Set<OfferEmploymentType> offerEmploymentTypes = new HashSet<>();
+    @OneToMany(mappedBy = "technology", cascade = CascadeType.ALL, orphanRemoval = true)
+    private Set<OfferTechnology> offerTechnologies = new HashSet<>();
+
 }
