+++ jobportal/jobportal/src/main/java/com/jobportal/jobportal/repositories/OfferRepository.java
@@ -1,0 +1,10 @@
+package com.jobportal.jobportal.repositories;
+
+import com.jobportal.jobportal.entities.offer.Offer;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OfferRepository extends JpaRepository<Offer, Long> {
+
+}
