+++ jobportal/jobportal/src/main/java/com/jobportal/jobportal/services/OfferServiceImpl.java
@@ -1,0 +1,4 @@
+package com.jobportal.jobportal.services;
+
+public class OfferServiceImpl {
+}
